Fix get userId on reload
